+++ src/test/scala/com.gu.contentapi.client/GuardianContentClientTest.scala
@@ -1,20 +1,18 @@
 package com.gu.contentapi.client
 
-import com.gu.contentapi.client.model.{TagsQuery, CollectionQuery, ItemQuery}
+import com.gu.contentapi.client.model.{CollectionQuery, ItemQuery}
 import org.joda.time.DateTime
 import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.time.{Seconds, Span, Millis}
+import org.scalatest.time.{Seconds, Span}
 import org.scalatest.{OptionValues, FlatSpec, Matchers}
 
-import scala.concurrent.ExecutionContext
+import scala.concurrent.ExecutionContext.Implicits.global
 
 class GuardianContentClientTest extends FlatSpec with Matchers with ClientTest with ScalaFutures with OptionValues {
-
-  implicit def executionContext = ExecutionContext.global
   implicit override val patienceConfig = PatienceConfig(timeout = Span(2, Seconds))
 
   "client interface" should "successfully call the Content API" in {
-    val query = ItemQuery().itemId("commentisfree/2012/aug/01/cyclists-like-pedestrians-must-get-angry")
+    val query = ItemQuery("commentisfree/2012/aug/01/cyclists-like-pedestrians-must-get-angry")
     val content = for {
       response <- api.getResponse(query)
     } yield response.content.get
@@ -22,7 +20,7 @@ class GuardianContentClientTest extends FlatSpec with Matchers with ClientTest w
   }
 
   it should "return errors as a broken promise" in {
-    val query = ItemQuery().itemId("something-that-does-not-exist")
+    val query = ItemQuery("something-that-does-not-exist")
     val errorTest = api.getResponse(query) recover { case error =>
       error should be (GuardianContentApiError(404, "Not Found"))
     }
@@ -49,13 +47,13 @@ class GuardianContentClientTest extends FlatSpec with Matchers with ClientTest w
   }
 
   it should "perform a given item query" in {
-    val query = ItemQuery().itemId("commentisfree/2012/aug/01/cyclists-like-pedestrians-must-get-angry")
+    val query = ItemQuery("commentisfree/2012/aug/01/cyclists-like-pedestrians-must-get-angry")
     val content = for (response <- api.getResponse(query)) yield response.content.get
     content.futureValue.id should be ("commentisfree/2012/aug/01/cyclists-like-pedestrians-must-get-angry")
   }
 
   it should "perform a given collection query" in {
-    val query = CollectionQuery().collectionId("uk-alpha/news/regular-stories")
+    val query = CollectionQuery("uk-alpha/news/regular-stories")
     val collection = for (response <- api.getResponse(query)) yield response.collection
     collection.futureValue.id should equal("uk-alpha/news/regular-stories")
   }